requirements.txt
ipykernel
requests
tensorflow
pandas 
numpy 
seaborn
scikit-learn
tensorboard
matplotlib
scikeras
pillow
torch
torchvision
nltk
psycopg2 
flask
fastapi
uvicorn
pydantic
ultralytics
python-multipart
sqlalchemy
llama-index
python-dotenv
llama-index-core
llama-index-llms-ollama
llama-index-readers-database
ollama
langchain
langchain-community
langchain-openai
langchain-ollama
ipywidgets
transformers
opencv-python
deepface
aiosmtplib